+++ src-ui/pages/memory_updater/SlowLook.svelte
@@ -1,10 +1,8 @@
 <script type="ts">
   import Button, { Label } from "@smui/button";
-  import Card, { Content } from "@smui/card";
-  import Switch from "@smui/switch/src/Switch.svelte";
+  import Card from "@smui/card";
   import { derived } from "svelte/store";
 
-  import { autoFixerHeader, enabledAutoFixer } from "@hdt/config";
   import { memoryUpdaterData } from "@hdt/tasks";
   import { invoke } from "@tauri-apps/api/tauri";
 
@@ -59,23 +57,6 @@
       <Label>Fix slow look!</Label>
     </Button>
   </div>
-  <div>
-    <div class="auto-fixer-header">
-      <h3>Auto-Fixer</h3>
-      <Switch bind:checked={$enabledAutoFixer} />
-      <div
-        class="switch-state"
-        class:disabled={!$enabledAutoFixer}
-        class:enabled={$enabledAutoFixer}
-      >
-        {$autoFixerHeader}
-      </div>
-    </div>
-    <p>
-      The Slow Look auto-fixer will ensure your camera speed is always set to
-      the correct default between runs.
-    </p>
-  </div>
 </div>
 
 <style>
@@ -84,23 +65,6 @@
   }
 
   .card-container.needs-fix {
-    color: rgb(255, 73, 73);
-  }
-
-  div.auto-fixer-header > h3 {
-    border-right: 1px solid white;
-    padding-right: 25px;
-    margin-right: 10px;
-  }
-  .auto-fixer-header {
-    display: flex;
-    align-items: center;
-  }
-  .switch-state.enabled {
-    color: rgb(18, 239, 18);
-  }
-
-  .switch-state.disabled {
     color: rgb(255, 73, 73);
   }