+++ src-ui/pages/memory_updater/config.ts
@@ -1,14 +1,18 @@
-import { writable } from "svelte/store";
+import { derived, writable } from "svelte/store";
 
 import { Store } from "tauri-plugin-store-api";
 
 export const configLoaded = writable(false);
 
 // Defaults
-export const showKills = writable(false);
+export const autoFixSlowLook = writable(true);
+export const autoFixCharacters = writable(false);
+export const desiredCharacters = writable([]);
 
 let configMapping = {
-  "show-kills": showKills,
+  "auto-fix-slow-look": autoFixSlowLook,
+  "auto-fix-characters": autoFixCharacters,
+  "desired-characters": desiredCharacters,
 };
 
 let subscriptions = {};
@@ -21,7 +25,7 @@ function setupSubscribers() {
   }
 }
 
-const tauriStore = new Store("tracker-pacifist.config");
+const tauriStore = new Store("auto-fixer.config");
 tauriStore
   .entries()
   .then((entries) => {